+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Update the logic to calculate price with multiplier units
+
+### Added
+- Test for prices with multiplier units
+
 ## [2.164.0] - 2023-06-12
 
 ### Fixed